requirements.txt
colorama==0.4.6
yaspin==2.1.0
newspaper3k==0.2.8
ollama==0.2.0
requests==2.31.0
